Refactor: 로그인, 아이템, 서비스 로직 변경
- 로그인 토큰 시간변경 (엑세스: 30분, 리프레시: 3일)
- 아이템 페이지 프롭스 내려주는것을 이제 명확히 리펙토링
- 서비스 로직 다시 만들어둔 default, cookie 패치 이용
- !! 인가되지않은 사용자 api 수정

+++ src/services/comment-service.js
@@ -1,54 +1,26 @@
-export async function createItemComment(id, formData) {
-  const res = await fetch(`/api/items/${id}/comment`, {
-    method: 'POST',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify(formData),
-  });
+import { cookieFetch } from './fetch-client';
 
-  if (!res.ok) {
-    const errorObj = await res
-      .json()
-      .catch(() => ({ message: '댓글 생성중 오류 발생' }));
-    throw new Error(`문의 댓글을 생성하지 못하였습니다.${errorObj.message}`);
-  }
+export const commentService = {
+  createComments: (id, formData) => {
+    if (!id) throw new Error('문의 댓글을 가져오지 못했습니다.');
+    cookieFetch(`/api/items/${id}/comment`, {
+      method: 'POST',
+      body: JSON.stringify(formData),
+    });
+  },
 
-  const result = await res.json();
-  return result.data;
-}
+  updateComments: (id, formData) => {
+    if (!id) throw new Error('아이템 아이디를 찾지못하였습니다.');
+    cookieFetch(`/api/items/${id}/comment`, {
+      method: 'PATCH',
+      body: JSON.stringify(formData),
+    });
+  },
 
-export async function updateItemComment(id, formData) {
-  const res = await fetch(`/api/items/${id}/comment`, {
-    method: 'PATCH',
-    headers: {
-      'Content-Type': 'application/json',
-    },
-    body: JSON.stringify(formData),
-  });
-  if (!res.ok) {
-    const errorObj = await res
-      .json()
-      .catch(() => ({ message: '댓글 업데이트중 오류' }));
-    throw new Error(`문의 댓글을 삭제하지 못하였습니다.${errorObj.message}`);
-  }
-
-  const result = await res.json();
-  return result.data;
-}
-
-export async function deleteItemComment(id, commentId) {
-  const res = await fetch(`/api/items/${id}/comment?commentId=${commentId}`, {
-    method: 'DELETE',
-  });
-
-  if (!res.ok) {
-    const errorObj = await res
-      .json()
-      .catch(() => ({ message: '댓글 삭제중 오류 발생' }));
-    throw new Error(`문의 댓글을 삭제하지 못하였습니다.${errorObj.message}`);
-  }
-
-  const result = await res.json();
-  return result.data;
-}
+  deleteComments: (id, commentId) => {
+    if (!id) throw new Error('상품아이디를 찾을수없습니다.');
+    cookieFetch(`/api/items/${id}/comment?commentId=${commentId}`, {
+      method: 'DELETE',
+    });
+  },
+};
